+++ mqttany/modules/gpio/GPIO/common.py
@@ -1,0 +1,49 @@
+"""
+***************************
+GPIO Library Wrapper Common
+***************************
+
+:Author: Michael Murton
+"""
+# Copyright (c) 2019 MQTTany contributors
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+from enum import Enum
+
+all = [ "Logic", "Direction", "Resistor", "Interrupt" ]
+
+class Logic(Enum):
+    LOW = 0
+    HIGH = 1
+
+class Direction(Enum):
+    INPUT = 10
+    OUTPUT = 11
+
+class Resistor(Enum):
+    OFF = 20
+    PULL_UP = 21
+    PULL_DOWN = 22
+
+class Interrupt(Enum):
+    NONE = 0
+    RISING = 30
+    FALLING = 31
+    BOTH = 32